pyproject.toml
[tool.poetry]
name = "restauant-picker"
version = "0.1.0"
description = "Randomly select a restaurant"
authors = ["Hugh Wimberly <hugh.wimberly@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
"aws-cdk.core" = "^1.91.0"
"aws-cdk.aws-s3" = "^1.91.0"
"aws-cdk.aws-iam" = "^1.91.0"
"aws-cdk.aws-lambda" = "^1.91.0"
"aws-cdk.aws-apigateway" = "^1.91.0"
"aws-cdk.aws-dynamodb" = "^1.91.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
isort = "^5.7.0"
pre-commit = "^2.10.1"
pytest = "^6.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
